4.21

+++ osa4/4.1-/controllers/blogs.js
@@ -48,8 +48,20 @@ const getTokenFrom = request => {
   })
 
   blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+    if (!decodedToken || !decodedToken.id) {
+      return response.status(401).json({error: 'token invalid'})
+    }
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (blog.user.toString() === decodedToken.id.toString()) {
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    } else {
+      response.status(401).end()
+    }
+    
   })
 
   blogsRouter.put('/:id', async (request, response, next) => {
